+++ crates/web-client/src/new_transactions.rs
@@ -43,11 +43,28 @@ impl WebClient {
         &mut self,
         transaction_result: &TransactionResult,
     ) -> Result<(), JsValue> {
+        let remote_prover = self.remote_prover.clone();
         if let Some(client) = self.get_mut_inner() {
             let native_transaction_result: NativeTransactionResult = transaction_result.into();
-            client.submit_transaction(native_transaction_result).await.map_err(|err| {
-                JsValue::from_str(&format!("Failed to submit Transaction: {}", err))
-            })?;
+            match remote_prover {
+                Some(ref remote_prover) => {
+                    client
+                        .submit_transaction_with_prover(
+                            remote_prover.clone(),
+                            native_transaction_result,
+                        )
+                        .await
+                        .map_err(|err| {
+                            JsValue::from_str(&format!("Failed to submit Transaction: {}", err))
+                        })?;
+                },
+                None => {
+                    client.submit_transaction(native_transaction_result).await.map_err(|err| {
+                        JsValue::from_str(&format!("Failed to submit Transaction: {}", err))
+                    })?;
+                },
+            }
+
             Ok(())
         } else {
             Err(JsValue::from_str("Client not initialized"))